feat: delete config
