+++ src/frontend/src/lib/components/app/core/Confirmation.svelte
@@ -9,16 +9,16 @@
 		visible?: boolean;
 		title?: Snippet;
 		children: Snippet;
-		size?: "default" | "wide"
+		size?: 'default' | 'wide';
 	}
 
-	let { visible = $bindable(false), title, children, size = "default" }: Props = $props();
+	let { visible = $bindable(false), title, children, size = 'default' }: Props = $props();
 
 	const dispatch = createEventDispatcher();
 	const yes = () => dispatch('junoYes');
 	const no = () => dispatch('junoNo');
 
-	let center = $derived(size === "wide" ? "wide" : true)
+	let center = $derived(size === 'wide' ? 'wide' : true);
 </script>
 
 <Popover backdrop="dark" {center} bind:visible>
