+++ src/frontend/src/lib/components/satellites/automation/repository/RepositoryConnectForm.svelte
@@ -5,7 +5,7 @@
 	import Collapsible from '$lib/components/ui/Collapsible.svelte';
 	import Input from '$lib/components/ui/Input.svelte';
 	import Value from '$lib/components/ui/Value.svelte';
-	import { buildRepositoryKey } from '$lib/services/satellite/automation/automation.config.services';
+	import { buildRepositoryKey } from '$lib/services/satellite/automation/repository.services';
 	import { i18n } from '$lib/stores/app/i18n.store';
 	import { toasts } from '$lib/stores/app/toasts.store';
