+++ src/frontend/src/lib/services/satellite/automation/automation.config.edit.services.ts
@@ -1,26 +1,14 @@
 import type { SatelliteDid } from '$declarations';
-import { setAutomationConfig } from '$lib/api/satellites.api';
 import { AUTOMATION_DEFAULT_MAX_SESSION_TIME_TO_LIVE } from '$lib/constants/automation.constants';
-import { loadSatelliteConfig } from '$lib/services/satellite/satellite-config.services';
-import { i18n } from '$lib/stores/app/i18n.store';
-import { toasts } from '$lib/stores/app/toasts.store';
+import {
+	type UpdateAutomationConfigResult,
+	updateConfig
+} from '$lib/services/satellite/automation/_automation.config.services';
 import type { AddAccessKeyScope } from '$lib/types/access-keys';
 import type { OptionIdentity } from '$lib/types/itentity';
 import type { Satellite } from '$lib/types/satellite';
 import type { WorkflowReferences } from '$lib/types/workflow';
-import { isNullish, toNullable } from '@dfinity/utils';
-import type { Identity } from '@icp-sdk/core/agent';
-import { get } from 'svelte/store';
-
-export interface UpdateAutomationConfigResult {
-	success: 'ok' | 'error';
-	err?: unknown;
-}
-
-interface UpdateResult {
-	result: 'success' | 'error';
-	err?: unknown;
-}
+import { toNullable } from '@dfinity/utils';
 
 export const updateAutomationKeysConfig = async ({
 	automationConfig,
@@ -37,13 +25,6 @@ export const updateAutomationKeysConfig = async ({
 	satellite: Satellite;
 	identity: OptionIdentity;
 }): Promise<UpdateAutomationConfigResult> => {
-	const labels = get(i18n);
-
-	if (isNullish(identity) || isNullish(identity?.getPrincipal())) {
-		toasts.error({ text: labels.core.not_logged_in });
-		return { success: 'error' };
-	}
-
 	const updateProviderConfig: SatelliteDid.OpenIdAutomationProviderConfig = {
 		...providerConfig,
 		controller:
@@ -81,13 +62,6 @@ export const updateAutomationConnectRepositoryConfig = async ({
 	satellite: Satellite;
 	identity: OptionIdentity;
 }): Promise<UpdateAutomationConfigResult> => {
-	const labels = get(i18n);
-
-	if (isNullish(identity) || isNullish(identity?.getPrincipal())) {
-		toasts.error({ text: labels.core.not_logged_in });
-		return { success: 'error' };
-	}
-
 	const updateProviderConfig: SatelliteDid.OpenIdAutomationProviderConfig = {
 		...providerConfig,
 		repositories: [...providerConfig.repositories, [repoKey, { refs: toNullable(repoReferences) }]]
@@ -99,77 +73,4 @@ export const updateAutomationConnectRepositoryConfig = async ({
 		automationConfig,
 		providerConfig: updateProviderConfig
 	});
-};
-
-const updateConfig = async ({
-	satellite,
-	automationConfig,
-	providerConfig,
-	identity
-}: {
-	satellite: Satellite;
-	automationConfig: SatelliteDid.AutomationConfig;
-	providerConfig: SatelliteDid.OpenIdAutomationProviderConfig;
-	identity: Identity;
-}): Promise<UpdateAutomationConfigResult> => {
-	const updateAutomationConfig: SatelliteDid.SetAutomationConfig = {
-		...automationConfig,
-		openid: toNullable({
-			observatory_id: [],
-			providers: [[{ GitHub: null }, providerConfig]]
-		})
-	};
-
-	const result = await setConfig({
-		identity,
-		satellite,
-		config: updateAutomationConfig
-	});
-
-	if (result.result === 'error') {
-		return {
-			success: 'error',
-			err: result.err
-		};
-	}
-
-	if (result.result === 'success') {
-		// Reload Satellite configuration
-		await loadSatelliteConfig({
-			identity,
-			satelliteId: satellite.satellite_id,
-			reload: true
-		});
-	}
-
-	return { success: 'ok' };
-};
-
-const setConfig = async ({
-	satellite: { satellite_id: satelliteId },
-	config,
-	identity
-}: {
-	satellite: Satellite;
-	config: SatelliteDid.SetAutomationConfig;
-	identity: Identity;
-}): Promise<UpdateResult> => {
-	try {
-		await setAutomationConfig({
-			satelliteId,
-			config,
-			identity
-		});
-	} catch (err: unknown) {
-		const labels = get(i18n);
-
-		toasts.error({
-			text: labels.errors.save_automation_config,
-			detail: err
-		});
-
-		return { result: 'error', err };
-	}
-
-	return { result: 'success' };
 };
